Unify ASGI handler, remove old code
- Rename hornbeam_asgi_loop to hornbeam_asgi
- Remove old ASGI handler code from hornbeam_handler.erl
- Use direct reply for responses with Content-Length
- Use chunked encoding for responses without Content-Length
- Fix empty body handling: check Transfer-Encoding too
- Simplify Python worker: no buffering, read channel directly

Removes ~900 lines, all 38 ASGI tests pass.
